new changes in the components structure
